+++ src/MailableManager.php
@@ -24,4 +24,21 @@ class MailableManager extends DefaultPluginManager
         $this->alterInfo('wmmailable_info');
         $this->setCacheBackend($cacheBackend, 'wmmailable_info_plugins');
     }
+
+    protected function findDefinitions()
+    {
+        $definitions = parent::findDefinitions();
+
+        foreach ($definitions as $id => $definition) {
+            // Trim leading backslashes if the class is used as id
+            if (class_exists($id)) {
+                $newId = ltrim($id, '\\');
+                $definition['id'] = $newId;
+                unset($definitions[$id]);
+                $definitions[$newId] = $definition;
+            }
+        }
+
+        return $definitions;
+    }
 }