+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2020-07-16
+### Added
+- Allow using class name as plugin id
+- Add documentation on theming mails
+
+### Changed
+- Use class name as plugin id when id is not specified
+
 ## [2.0.2] - 2020-05-14
 ### Fixed
 - Fix wrong return types